+++ src/core/managers/internal/NotificationsFeedManager/helpers/index.ts
@@ -1,0 +1,3 @@
+export { mapTransactionAssets } from './mapTransactionAssets';
+export { mapTransactionToListItem } from './mapTransactionToListItem';
+export { getHistoricalTransactions } from './getHistoricalTransactions';