+++ CHANGELOG.md
@@ -7,8 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- [Added notifications feed manager](https://github.com/multiversx/mx-sdk-dapp-core-ui/pull/112)
+- [Added notifications feed historical transactions mapping](https://github.com/multiversx/mx-sdk-dapp-core-ui/pull/114)
 - [Added sign message helper](https://github.com/multiversx/mx-sdk-dapp-core/pull/113)
+- [Added notifications feed manager](https://github.com/multiversx/mx-sdk-dapp-core-ui/pull/112)
 - [Fixed minor TODOs](https://github.com/multiversx/mx-sdk-dapp-core/pull/110)
 - [Fixed getWindowParentOrigin.test](https://github.com/multiversx/mx-sdk-dapp-core/pull/111)
 - [Updated managers to use internal folder](https://github.com/multiversx/mx-sdk-dapp-core/pull/108)